requirements.txt
<<<<<<< HEAD
git+git://github.com/kristallizer/django-1.4.2.git
argparse==1.2.1
constants==0.6.0
django-devserver==0.7.0
django-nose==1.2
drest==0.9.12
httplib2==0.9
mimeparse==0.1.3
nose==1.3.0
psycopg2==2.5.2
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.2.1
six==1.5.2
slumber==0.6.0
sqlparse==0.1.11
stevedore==0.15
wsgiref==0.1.2
